# Conflicts:
#	handlers/fileserver.go
#	handlers/publish_video_action.go
#	models/comment.go
#	models/user.go
#	models/video.go
#	router/routes.go
#	services/video/feed_video.go

+++ models/comment.go
@@ -2,7 +2,7 @@ package models
 
 type Comment struct {
 	Model
-	VideoID int64  `json:"video_id"`
+	VideoID uint   `json:"video_id"`
 	UserID  uint   `json:"user_id"`
 	Content string `json:"content"`
 }
